+++ app/models/project_favorite.rb
@@ -1,0 +1,2 @@
+class ProjectFavorite < ActiveRecord::Base
+end