+++ db/migrate/20130806130920_create_project_favorites.rb
@@ -1,0 +1,14 @@
+class CreateProjectFavorites < ActiveRecord::Migration
+  def change
+    create_table :project_favorites do |t|
+      t.integer :project_id
+      t.integer :user_id
+      t.boolean :favorite, default: false, null: false
+
+      t.timestamps
+    end
+
+    add_index :project_favorites, :project_id
+    add_index :project_favorites, :user_id
+  end
+end