+++ test/models/project_favorite_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class ProjectFavoriteTest < ActiveSupport::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end