+++ CHANGELOG.md
@@ -2,11 +2,12 @@
 
 ### Enhancements
 - **Gem Changes**
-  - Updated to contour 2.1.0.beta5
+  - Updated to contour 2.1.0.beta6
 - **Project Changes**
   - Project emails field was removed since it is no longer used
   - Project settings page removed and consolidated with project show page
   - Project splash now displays a number of recent valid entries made for a project
+  - Projects can be favorited which reorders them on the project splash page, and puts the first three as links into the menu bar
 
 ## 0.19.0 (August 1, 2013)